+++ components/admin/layers/table/LayersTable.js
@@ -67,7 +67,11 @@ class LayersTable extends React.Component {
           link={{
             label: 'New layer',
             route: 'admin_data_detail',
-            params: { tab: 'layers', id: 'new', dataset: this.props.dataset }
+            params: {
+              tab: 'layers',
+              id: 'new',
+              ...!!dataset && { dataset }
+            }
           }}
           onSearch={this.onSearch}
         />
@@ -76,7 +80,7 @@ class LayersTable extends React.Component {
         {!this.props.error && (
           <CustomTable
             columns={[
-              { label: 'Name', value: 'name', td: NameTD },
+              { label: 'Name', value: 'name', td: NameTD, tdProps: { dataset } },
               { label: 'Provider', value: 'provider' },
               { label: 'Updated at', value: 'updatedAt', td: UpdatedAtTD },
               { label: 'Ownership', value: 'userId', td: OwnershipTD, tdProps: { user } }
@@ -84,7 +88,7 @@ class LayersTable extends React.Component {
             actions={{
               show: true,
               list: [
-                { name: 'Edit', route: 'admin_data_detail', params: { tab: 'layers', subtab: 'edit', id: '{{id}}' }, show: true, component: EditAction },
+                { name: 'Edit', route: 'admin_data_detail', params: { tab: 'layers', subtab: 'edit', id: '{{id}}', ...!!dataset && { dataset } }, show: true, component: EditAction },
                 { name: 'Remove', route: 'admin_data_detail', params: { tab: 'layers', subtab: 'remove', id: '{{id}}' }, component: DeleteAction, componentProps: { authorization: this.props.authorization } },
                 { name: 'Go to dataset', route: 'admin_data_detail', params: { tab: 'datasets', subtab: 'edit', id: '{{id}}' }, component: GoToDatasetAction }
               ]