+++ redactions/index.js
@@ -24,6 +24,7 @@ export { default as widgets } from './admin/widgets';
 export { default as dashboards } from './admin/dashboards';
 export { default as partners } from './admin/partners';
 export { default as pages } from './admin/pages';
+export { default as sources } from './admin/sources';
 
 // Toastr
 export { reducer as toastr } from 'react-redux-toastr';