+++ CHANGELOG.md
@@ -8,9 +8,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.1.0] - 2017-10-11
 
 ### Fixed
+ - Planet pulse dropdowns have now a z-index that makes them be displayed on top of the layer card
 
 ### Added
-- Implementation of decision tree to get visualisations by dataset. [PDF](https://vizzuality.slack.com/files/U4C2Q99RB/F79L7J7UL/rw_-_widget_creation.pdf)
+- New prototype of tags tooltip in Explore
+- New search input for the filter tooltip in the Widget Editor
+- Implementation of decision tree to get visualizations by dataset. [PDF](https://vizzuality.slack.com/files/U4C2Q99RB/F79L7J7UL/rw_-_widget_creation.pdf)
 - New command to check datasets for widget editor `yarn run check-datasets`
 - Vega Chart as component
 - Added Vega Chart Thumbnail with theme configuration